Render flags
I encountered some difficulty when trying to wipe the free the flag
nodes and re-render them. Initially I was checking the node name to see
if it was "Flag". I played around with reversing the array of nodes,
but this never quite worked right. After switching to check if the class
is "Sprite2D", it worked fine. If the information panel had another
"Sprite2D", this would be problematic, but it would probably be best to
refactor this into a separate node with an associated script handling
this logic